Address DIK review comments
